honeypot/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# LLM
langchain-groq>=0.1.0

# Database & Cache (V2 Lite)
redis>=5.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
sqlalchemy[asyncio]>=2.0.0

# Background Tasks
celery[redis]>=5.3.0

# Utils
aiohttp>=3.9.0
networkx>=3.2.0

# ML Packages (OPTIONAL - commented for Lite version)
# Uncomment below for full ML features:
# sentence-transformers>=2.2.0
# spacy>=3.7.0
# numpy>=1.24.0
# geoip2>=4.8.0
# python-whois>=0.8.0
# weasyprint>=60.0
# jinja2>=3.1.0
